Adds network call when Ballot and VoteForm are rendered
