+++ src/routes/new.jsx
@@ -81,6 +81,7 @@ export default function NewBallot() {
             <div className='response-field'>
                 <span className='new-form-label'>Time limit in minutes</span>
                 <select name="timer" className='new-ballot-inp'>
+                    <option value="1">1 minute DEBUG</option>
                     <option value="5">5 minutes</option>
                     <option value="10">10 minutes</option>
                     <option value="20">20 minutes</option>
